+++ src/main/java/roomescape/dto/reservation/request/UserReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto.reservation;
+package roomescape.dto.reservation.request;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
@@ -10,7 +10,7 @@ import roomescape.domain.time.Time;
 
 import java.time.LocalDate;
 
-public record ReservationRequest(
+public record UserReservationRequest(
         @NotNull @DateTimeFormat(pattern = "yyyy-MM-dd") LocalDate date,
         @NotNull @Positive Long timeId,
         @NotNull @Positive Long themeId
