+++ src/main/java/roomescape/dto/reservation/response/ReservationAvailableTimeResponse.java
@@ -1,6 +1,7 @@
-package roomescape.dto.reservation;
+package roomescape.dto.reservation.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
+
 import java.time.LocalTime;
 
 public record ReservationAvailableTimeResponse(
