+++ src/main/java/roomescape/dto/reservation/response/ReservationResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto.reservation;
+package roomescape.dto.reservation.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import roomescape.domain.reservation.Reservation;
